new archi and add userDetail page, wip
