+++ src/Components/task/activeTaskStyle.css
@@ -1,0 +1,3 @@
+.card{
+    min-width: 370px;
+}
