+++ src/Constants/routeName.jsx
@@ -1,0 +1,3 @@
+export const HOME_PAGE_URL = '';
+export const TASKS_PAGE_URL = 'tasks';
+export const USERS_PAGE_URL = 'users';
