+++ src/Containers/home/homeContainer.jsx
@@ -1,0 +1,11 @@
+import React from "react"
+
+class HomeContainer extends React.Component {
+    render(){
+        return(
+            <p>home container</p>
+        )
+    }
+}
+
+export default HomeContainer;
