+++ src/Stylesheets/app.css
@@ -1,0 +1,27 @@
+.root {
+  display: flex;
+}
+.hide{
+  display: none!important;
+}
+.toolbar{
+ padding-right: 24px;
+}
+.menuButton{
+  margin-left: 12px;
+  margin-right: 36px;
+}
+.grow{
+  flex-grow: 1;
+}
+.flexCenter {
+  display: flex;
+  justify-content: center;
+}
+.flexEnd {
+  display: flex;
+  justify-content: flex-end;
+}
+.padding4{
+  padding: 4px;
+}
