+++ src/Stylesheets/themeConfiguration.jsx
@@ -1,0 +1,8 @@
+import { teal } from "@material-ui/core/colors";
+import createMuiTheme from "@material-ui/core/styles/createMuiTheme";
+import createPalette from "@material-ui/core/styles/createPalette";
+export const muiTheme = createMuiTheme({
+    palette: createPalette({
+        primary: teal,
+    }),
+});
